backend/comment_service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
pydantic>=2.4.0
motor>=3.3.1
python-dotenv>=1.0.0
httpx>=0.25.0
python-multipart>=0.0.6
PyJWT>=2.8.0